deepen: inline openrouter-credit-monitor into daily-report-handler
Replaces `lib/ai/openrouter-credit-monitor.ts` (160 LOC, 3 public
exports: `getOpenRouterCreditStatus`, `formatCreditStatusForSlack`,
`CreditStatus`) with private helpers of the same names inside its
sole caller, `lib/slack/daily-report-handler.ts`.

Why this is a deepening
-----------------------
The source module already declared in its docstring: "The only
production caller is `lib/slack/daily-report-handler.ts`." That's the
textbook one-adapter/no-real-seam shape. The interface was almost as
wide as the body: one `fetch()` to /api/v1/auth/key plus one string-
templating switch statement, split behind a public interface used by
exactly one call site (`generateHourlySummary`) and one downstream
formatter site (`formatIntervalSummaryMessage`).

Deletion test: fold the two functions and the `CreditStatus`
interface back into `daily-report-handler.ts`. Nothing spreads across
N callers because there are no N callers. Same shape as ADR-0002
and the CONTEXT.md "Why It Matters Guard" / "Ticker Grounding Guard"
precedents. Follows the "-monitor.ts wraps one fetch + one string
templater with 1 caller" pattern the routine has been consolidating.

Dependency category
-------------------
True-external (OpenRouter's REST endpoint), but with exactly one
caller. Per LANGUAGE.md, "one adapter means a hypothetical seam" —
there is no second adapter that would justify the port. The Slack
report is the only surface that reads credit status; a future second
caller can extract a port at that time.

Tests
-----
No test files referenced `lib/ai/openrouter-credit-monitor.ts`
(confirmed by grep before deletion). The Slack report's behaviour is
exercised through its existing `daily-report-handler.ts` test surface.
No test coverage lost.

Files
-----
- `lib/ai/openrouter-credit-monitor.ts` — deleted.
- `lib/slack/daily-report-handler.ts` — private
  `getOpenRouterCreditStatus` / `formatCreditStatusForSlack` /
  `CreditStatus` helpers added at the top of the file; two existing
  call sites (`generateHourlySummary`, the credit-block render in
  `formatIntervalSummaryMessage` context) already reference them by
  bare name post-inline.
